+++ partitioning.py
@@ -19,7 +19,7 @@ PERSON_COLUMN = 1
 def get_persons_ids():
     return np.unique(identity_df[PERSON_COLUMN])
 
-if __name__ == '__main__':
+def build_celeba_partitions():
     ids = get_persons_ids()
     np.random.shuffle(ids)
     
@@ -30,5 +30,11 @@ if __name__ == '__main__':
     if not os.path.exists('celeba_partitions'):
         os.mkdir('celeba_partitions')
 
-    with open(os.path.join('celeba_partitions', 'partitions.json'), 'w') as file:
-        json.dump({'eval': [int(x) for x in eval_ids], 'training': [int(x) for x in training_ids]}, file, indent=4)
+    if not os.path.exists(os.path.join('celeba_partitions', 'partitions.json')):
+        with open(os.path.join('celeba_partitions', 'partitions.json'), 'w') as file:
+            json.dump({'eval': [int(x) for x in eval_ids], 'training': [int(x) for x in training_ids]}, file, indent=4)
+    else:
+        print('Partitions file already generated')
+
+if __name__ == '__main__':
+    build_celeba_partitions()
